Test Passed - There is validation for if a reservation is cancelled, but there wasn't one made initially.

+++ src/main/java/library/ReservationService.java
@@ -45,13 +45,15 @@ public class ReservationService {
      */
     public void cancel(String userId, String bookId) {
         // TODO: Implement using TDD
+        if (!reservationRepo.existsByUserAndBook(userId, bookId)) {
+            throw new IllegalArgumentException("No reservations found");
+        }
+
         reservationRepo.delete(userId, bookId);
 
         Book book = bookRepo.findById(bookId);
         book.setCopiesAvailable(book.getCopiesAvailable()+  1);
         bookRepo.save(book);
-
-
     }
 
     /**
